parties

cowcode

Country code (from Correlates of War Project).

ccode

Country code (International Standards Organization, Alpha-3 code)

year

Year

legis

Legislature in operation?

date_low

Date of the election for the Lower House

date_upp

Date of the election for the Upper House

pty_acrn

Party acronym

pty_name

Party name (full party name, as reported by Political Handbook of the World or by your other sources).

pty_code

Nine-digit party code: The first three digits match the country code; the next four digits are the year in which the party was founded; the last two digits identify the specific party, in case more than one party was created in the same country and the same year. For example, the Radical Party (a.k.a. Uni?n C?vica Radical or UCR) was founded in Argentina (cowcode = 160) in 1890. Thus, the code for the party is 160189001.

fac_name

Name of the faction. Factions may lack a formal name; in that case, is added a brief note to identify the faction (e.g., "Dissident Liberals"). If the row represents a whole party (which is the norm) the variable is left blank.

fac_code

Twelve-digit faction code: The first nine digits are equal to the party code (pty_code); the last three digits are an arbitrary sequence to identify the specific faction within the party (e.g., 001).

s_low

Number of seats obtained by the party (faction) in the Lower (only) House. If there is no legislature, or if the party has no seats in the lower House, the variable is left blank.

ts_low

Total number of seats in the Lower (or only) House (including vacant seats). If no legislature, it is left blank.

s_upp

Number of seats obtained by the party (faction) in the Upper House. Note: If there is no Senate, or if the party has no seats in the Senate, it is left blank.

ts_upp

Total number of seats in the Upper House (including vacant seats). Note: If no legislature, it is left blank.

presp

Coded 1 if the line corresponds to the president's party; 0 otherwise.

cl_other

Coded 1 if party belongs to the president's coalition; 0 otherwise. Coalitions are blocs that consistently support the president in the legislature. Different sources are used to define which parties (or factions) supported the president.

cl_altman

Coalitions coded as is in: Altman, David. 2001. The Politics of Coalition Formation and Survival in Multiparty Presidential Regimes. Ph.D. Dissertation, University of Notre Dame.

cl_Deheza

Coalitions coded as is in: Deheza, Grace Ivana. 1997. Gobiernos de coalici?n en el sistema presidencial: America del Sur, Departamento de Ciencias Pol?ticas y Sociales, Instituto Universitario Europeo, Florencia.

cl_DPI

Coalitions coded as is in: The Database of Political Institutions (DPI). Available in: https://publications.iadb.org/handle/11319/8806

founded

Year in which the party was founded. We use 8888 as a default.

source

Source for the information reported.

data(parties)

Format

A data frame with ... rows and ... variables

Source

An?bal P?rez-Li??n and Nicol?s Schmidt and Daniela Vairo (2019), Presidential hegemony and democratic backsliding in Latin America, 1925-2016, Democratization, 26, 4, pp.606-625, https://doi.org/10.1080/13510347.2019.1566321.